Add support for multi-speed pumps
Modified generateFanAdJSON so it can be leveraged by the blower and multi-speed pumps.

Single speed pumps still use a switch, but to enable them to use a common state callback,
I had to change the propertyId to have the suffix "_state". This means new pump entities are added.
This can be avoided by having more pump state options in the MQTT callback.

We should talk about the best approach for the single speed pumps.