Added mDNS support
This enables the eSpa to be accessed via the spa name (defaults to MySpa)
instead of needing to know the IP address: myspa.local

The Spa Name is configurable, therefore this change necessitated moving
the config read above the Wi-Fi connection in setup().
I don't see any reason not to do this.

I think we should change the default spa name to eSpa to match the new
project name. This change will need to be done carefully, as it could
result in a breaking change.